WEP-50 move header to layouts
