+++ frontend/src/features/auth/index.ts
@@ -1,4 +1,3 @@
 export { useAuthStore } from './store/auth';
 export { authService } from './api/auth.service';
 export { default as UserIdentity } from './components/UserIdentity.vue';
-// Add other public components or logic here as needed
