+++ frontend/src/features/week/index.ts
@@ -2,4 +2,3 @@ export { useWeekStore } from './store/weeks';
 export { weekService } from './api/week.service';
 export { default as WeekDetails } from './components/WeekDetails.vue';
 export { default as WeekCreateForm } from './components/WeekCreateForm.vue';
-// Add other public components or logic here as needed
