+++ frontend/src/layouts/TheGuestLayout.vue
@@ -19,6 +19,6 @@
 </template>
 
 <script setup lang="ts">
-import GuestAppHeader from '@/components/header/guest/GuestAppHeader.vue';
 import GuestFooter from '@/layouts/components/GuestFooter.vue';
+import GuestAppHeader from './components/header/guest/GuestAppHeader.vue';
 </script>
