+++ frontend/src/layouts/components/header/auth/AuthAppHeader.spec.ts
