+++ frontend/src/layouts/components/header/auth/AuthAppHeader.vue
@@ -20,7 +20,7 @@
 <script setup lang="ts">
 import { SidebarTrigger } from '@/components/ui/sidebar';
 
-import AppHeader from '@/components/header/AppHeader.vue';
+import AppHeader from '../AppHeader.vue';
 import ModeToggle from '@/components/shared/ModeToggle.vue';
 import Separator from '@/components/ui/separator/Separator.vue';
