+++ frontend/src/layouts/components/header/guest/GuestAuthActions.spec.ts
