+++ frontend/src/layouts/components/header/guest/GuestAuthActions.vue
