+++ frontend/src/layouts/components/header/guest/GuestMobileMenu.spec.ts
